Refactor CommandSender class
