+++ src/main/java/ru/sbt/mipt/oop/scenarios/ScenarioController.java
@@ -2,10 +2,9 @@ package ru.sbt.mipt.oop.scenarios;
 
 import ru.sbt.mipt.oop.event.CommandType;
 import ru.sbt.mipt.oop.event.SensorCommand;
-import ru.sbt.mipt.oop.event.controllers.HomeCommandController;
+import ru.sbt.mipt.oop.event.CommandSender;
 import ru.sbt.mipt.oop.smarthome.Light;
 import ru.sbt.mipt.oop.smarthome.LightIterator;
-import ru.sbt.mipt.oop.smarthome.Room;
 import ru.sbt.mipt.oop.smarthome.SmartHome;
 
 // Интерфейс обработчика сценариев умного дома
@@ -29,7 +28,7 @@ public class ScenarioController {
             light.setOn(false);
             SensorCommand command = new SensorCommand(CommandType.LIGHT_OFF, light.getId());
             System.out.println("Light " + light.getId() + " was turned off by HallDoor.");
-            HomeCommandController.sendCommand(command);
+            CommandSender.sendCommand(command);
         }
     }
 }
